Modified game.cpp

+++ exam-project/arduino/sketch/slotmachine.h
@@ -30,7 +30,7 @@ US dollars to credit conversion, price per spin etc.
 typedef struct slotmachine_config
 {
     int spin_credit_price;
-    float usd_to_credit_conversion;
+    int usd_to_credit_conversion;
 } slotmachine_config_t;
 
 /*
@@ -74,7 +74,7 @@ the pointer.
 */
 
 slotmachine_t* create_slotmachine (slotmachine_config_t config, wheel_t* wheels[], rule_t rules[], size_t rules_size);
-slotmachine_config_t create_slotmachine_config (int spin_credit_price, float usd_credit_conversion);
+slotmachine_config_t create_slotmachine_config (int spin_credit_price, int usd_credit_conversion);
 wheel_t* create_wheel (symbol_t symbols[]);
 rule_t create_rule (symbol_t symbols[], int price);
 
@@ -82,7 +82,7 @@ rule_t create_rule (symbol_t symbols[], int price);
 Creates a slotmachine struct with hardcoded values specific to the
 assignment requirements. See README for more information.
 */
-slotmachine_t* create_default_slotmachine ();
+slotmachine_t* create_default_slotmachine();
 
 /*
 Function which deallocates the allocated memory for the slotmachine.
